some refactore + declared new routes

+++ reactapp/src/features/dashboard/Dashboard.tsx
@@ -2,32 +2,63 @@ import { Grid, SemanticWIDTHS } from "semantic-ui-react";
 import { useStore } from "../../app/stores/store";
 import { observer } from "mobx-react-lite";
 import NoteList from "../notes/list/NoteList";
+import AddNoteButton from "../notes/other/AddNoteButton";
+import ColumnButton from "../notes/other/ColumnButton";
+import { useEffect } from "react";
+import LoadingComponent from "../../app/layout/LoadingComponent";
 
 function Dashboard() {
-    const { pageStore, noteStore } = useStore();
+    const { pageStore, noteStore, commonStore, userStore } = useStore();
     const { selectedElement, editMode } = pageStore;
 
+    useEffect(() => {
+        if (commonStore.token) {
+            userStore.getUser().finally(() => commonStore.setAppLoaded())
+        } else {
+            commonStore.setAppLoaded()
+        }
+    }, [commonStore, userStore])
+
+    if (!commonStore.appLoaded) return <LoadingComponent content="Loading app...." />
+
     return (
-        <Grid>
-            <Grid.Column width={16}>
-                {selectedElement && !editMode &&
-                    <>
-                        <NoteList
-                            noteArray={noteStore.getArray}
-                            noteLoading={noteStore.loading}
-                            noteEditMode={noteStore.editMode}
-                            noteSelectedElement={noteStore.selectedElement}
-                            noteOpenForm={noteStore.openForm}
-                            noteSelect={noteStore.selectOne}
-                            noteUpdate={noteStore.updateOne}
-                            noteDelete={noteStore.deleteOne}
-                            getNote={noteStore.getOne}
-                            columnsCount={noteStore.columnsCount as SemanticWIDTHS}
-                        />
-                    </>
+        <div id="main">
+            <div style={{ marginTop: '4em' }}>
+                {pageStore.selectedElement &&
+                    <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
+                        <div>
+                            <AddNoteButton />
+                        </div>
+                        <div>
+                            <ColumnButton />
+                        </div>
+                    </div>
                 }
-            </Grid.Column>
-        </Grid>
+            </div>
+            <div style={{ marginTop: '2em' }}>
+                <Grid>
+                    <Grid.Column width={16}>
+                        {selectedElement && !editMode &&
+                            <>
+                                <NoteList
+                                    noteArray={noteStore.getArray}
+                                    noteLoading={noteStore.loading}
+                                    noteEditMode={noteStore.editMode}
+                                    noteSelectedElement={noteStore.selectedElement}
+                                    noteOpenForm={noteStore.openForm}
+                                    noteSelect={noteStore.selectOne}
+                                    noteUpdate={noteStore.updateOne}
+                                    noteDelete={noteStore.deleteOne}
+                                    getNote={noteStore.getOne}
+                                    columnsCount={noteStore.columnsCount as SemanticWIDTHS}
+                                />
+                            </>
+                        }
+                    </Grid.Column>
+                </Grid>
+            </div>
+
+        </div>
     );
 }
